Enable Black and PyCodeStyle
