+++ .github/workflows/ci.yml
@@ -18,7 +18,7 @@ jobs:
           pip install -e .[dev]
       - name: Run ISort check
         run: python -m isort --check .
-      #- name: Run PyCodeStyle check
-      #  run: python -m pycodestyle .
-      #- name: Run Black check
-      #  run: python -m black --check .
+      - name: Run Black check
+       run: python -m black --check .
+      - name: Run PyCodeStyle check
+       run: python -m pycodestyle .
