+++ tests/buffers/test_buffer_factory.py
@@ -1,4 +1,3 @@
-import numpy
 import pytest
 
 from ai_traineree.buffers.buffer_factory import BufferFactory
@@ -7,28 +6,7 @@ from ai_traineree.buffers.per import PERBuffer
 from ai_traineree.buffers.replay import ReplayBuffer
 from ai_traineree.buffers.rollout import RolloutBuffer
 from ai_traineree.types.experience import Experience
-
-
-def generate_sample_SARS(iterations, obs_size: int = 4, action_size: int = 2, dict_type=False):
-    state_fn = lambda: numpy.random.random(obs_size)
-    action_fn = lambda: numpy.random.random(action_size)
-    reward_fn = lambda: float(numpy.random.random() - 0.5)
-    done_fn = lambda: numpy.random.random() > 0.5
-    state = state_fn()
-
-    for _ in range(iterations):
-        next_state = state_fn()
-        if dict_type:
-            yield dict(
-                state=list(state),
-                action=list(action_fn()),
-                reward=[reward_fn()],
-                next_state=list(next_state),
-                done=[bool(done_fn())],
-            )
-        else:
-            yield (list(state), list(action_fn()), reward_fn(), list(next_state), bool(done_fn()))
-        state = next_state
+from tests.buffers.test_replay import generate_sample_SARS
 
 
 def populate_buffer(buffer, num_samples):
