+++ tests/buffers/test_replay.py
@@ -1,29 +1,11 @@
 import copy
 
-import numpy
 import pytest
 
 from ai_traineree.buffers.replay import ReplayBuffer
 from ai_traineree.types.experience import Experience
 from ai_traineree.types.state import BufferState
-
-
-def generate_sample_SARS(iterations, obs_size: int = 4, action_size: int = 2, dict_type=False):
-    state_fn = lambda: list(numpy.random.random(obs_size))
-    action_fn = lambda: list(numpy.random.random(action_size))
-    reward_fn = lambda: float(numpy.random.random() - 0.5)
-    done_fn = lambda: numpy.random.random() > 0.5
-    state = state_fn()
-
-    for _ in range(iterations):
-        next_state = state_fn()
-        if dict_type:
-            yield dict(
-                state=state, action=action_fn(), reward=[reward_fn()], next_state=next_state, done=[bool(done_fn())]
-            )
-        else:
-            yield (state, action_fn(), reward_fn(), next_state, bool(done_fn()))
-        state = next_state
+from tests.utils import generate_sample_SARS
 
 
 def populate_buffer(buffer, num_samples):
